+++ pext/qml/InstallThemeFromURLDialog.qml
@@ -37,6 +37,16 @@ Dialog {
             id: textfield
             Layout.fillWidth: true
         }
+
+        Label {
+            text: qsTr("Enter the correct theme identifier:")
+        }
+
+        TextField {
+            id: identifierTextfield
+            Layout.fillWidth: true
+        }
+
     }
 
     Component.onCompleted: {
@@ -45,7 +55,7 @@ Dialog {
     }
 
     onAccepted: {
-        installRequest(textfield.text);
+        installRequest(textfield.text, identifierTextfield.text);
     }
 }